prompt-template-example/requirements.txt
flask
python-dotenv
requests
openai
numpy==1.24.2
aiofiles==23.1.0
